Adds routing to Pools api to access pool content and status

+++ src/pools/test/NIST.CVP.Pools.Tests/PoolTests.cs
@@ -1,9 +1,12 @@
-﻿using NIST.CVP.Common.Oracle.ParameterTypes;
+﻿using Newtonsoft.Json;
+using NIST.CVP.Common.Oracle.ParameterTypes;
 using NIST.CVP.Common.Oracle.ResultTypes;
 using NIST.CVP.Crypto.Common.Symmetric.Enums;
+using NIST.CVP.Generation.Core.JsonConverters;
 using NIST.CVP.Tests.Core;
 using NUnit.Framework;
 using System;
+using System.Collections.Generic;
 using System.IO;
 
 namespace NIST.CVP.Pools.Tests
@@ -14,6 +17,13 @@ namespace NIST.CVP.Pools.Tests
         private string _testPath;
         private string _configFile = "aes-test1.json";
         private string _fullPath;
+
+        private readonly IList<JsonConverter> _jsonConverters = new List<JsonConverter>
+        {
+            new BitstringConverter(),
+            new DomainConverter(),
+            new BigIntegerConverter()
+        };
 
         [SetUp]
         public void SetUp()
@@ -33,7 +43,7 @@ namespace NIST.CVP.Pools.Tests
                 KeyLength = 128
             };
 
-            var pool = new Pool<AesParameters, AesResult>(param, _fullPath);
+            var pool = new Pool<AesParameters, AesResult>(param, _fullPath, _jsonConverters);
 
             Assert.AreEqual(2, pool.WaterLevel);
         }
@@ -49,7 +59,7 @@ namespace NIST.CVP.Pools.Tests
                 KeyLength = 128
             };
 
-            var pool = new Pool<AesParameters, AesResult>(param, _fullPath);
+            var pool = new Pool<AesParameters, AesResult>(param, _fullPath, _jsonConverters);
 
             var result = pool.GetNext();
             
@@ -68,7 +78,7 @@ namespace NIST.CVP.Pools.Tests
                 KeyLength = 128
             };
 
-            var pool = new Pool<AesParameters, AesResult>(param, _fullPath);
+            var pool = new Pool<AesParameters, AesResult>(param, _fullPath, _jsonConverters);
 
             var result1 = pool.GetNext();
             var result2 = pool.GetNext();
@@ -91,13 +101,20 @@ namespace NIST.CVP.Pools.Tests
                 KeyLength = 128
             };
 
-            var pool = new Pool<AesParameters, AesResult>(param, _fullPath);
+            var pool = new Pool<AesParameters, AesResult>(param, _fullPath, _jsonConverters);
 
             var writePath = Path.Combine(_testPath, $"saveTest-{Guid.NewGuid().ToString().Substring(0, 8)}.json");
             pool.SavePoolToFile(writePath);
 
-            Assert.IsTrue(File.Exists(writePath));
-            File.Delete(writePath);
+            if (File.Exists(writePath))
+            {
+                File.Delete(writePath);
+                Assert.Pass();
+            }
+            else
+            {
+                Assert.Fail();
+            }
         }
     }
 }
